Disable "maybe-uninitialized" warning globally
Several build configurations had already disabled this warning
because it generates a lot of false positives. But some had not,
and it was still enabled for "allmodconfig" builds, for example.
Looking at the warnings produced, every single one I looked at
was a false positive, and the warnings are frequent enough
(and big enough) that they can easily hide real problems that
you don't notice in the noise generated by -Wmaybe-uninitialized.
The warning is good in theory, but this is a classic case of a
warning that causes more problems than the warning can solve.
If gcc gets better at avoiding false positives, we may be able
to re-enable this warning. But as is, we're better off without it,
and I want to be able to see the *real* warnings.

This fixes kernel building on stock 4.9 toolchains

Observed within commit for CM here
https://review.cyanogenmod.org/#/c/163331/

Thanks to @javelinanddart for reminding me it was
infact @zakooz who highlighted this commit. Thanks
Change-Id: I6ad1d4a0da0181a383290d02119455b8cdd87c9b